+++ activestorage/lib/active_storage/service/disk_service.rb
@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 require "fileutils"
 require "pathname"
 require "digest/md5"
 require "active_support/core_ext/numeric/bytes"
 
 module ActiveStorage
-  # Wraps a local disk path as a Active Storage service. See `ActiveStorage::Service` for the generic API
+  # Wraps a local disk path as a Active Storage service. See ActiveStorage::Service for the generic API
   # documentation that applies to all services.
   class Service::DiskService < Service
     attr_reader :root
@@ -124,4 +126,3 @@ module ActiveStorage
       end
   end
 end
-